+++ packages/@expo/cli/src/start/server/DevelopmentSession.ts
@@ -1,7 +1,10 @@
 import { ExpoConfig, getConfig } from '@expo/config';
 
 import { APISettings } from '../../api/settings';
-import { updateDevelopmentSessionAsync } from '../../api/updateDevelopmentSession';
+import {
+  closeDevelopmentSessionAsync,
+  updateDevelopmentSessionAsync,
+} from '../../api/updateDevelopmentSession';
 import { getUserAsync } from '../../api/user/user';
 import * as ProjectDevices from '../project/devices';
 
@@ -40,14 +43,14 @@ export class DevelopmentSession {
     runtime: 'native' | 'web';
   }): Promise<void> {
     if (APISettings.isOffline) {
-      this.stop();
+      this.stopNotifying();
       return;
     }
 
     const deviceIds = await this.getDeviceInstallationIdsAsync();
 
     if (!(await isAuthenticatedAsync()) && !deviceIds?.length) {
-      this.stop();
+      this.stopNotifying();
       return;
     }
 
@@ -60,7 +63,7 @@ export class DevelopmentSession {
       });
     }
 
-    this.stop();
+    this.stopNotifying();
 
     this.timeout = setTimeout(() => this.startAsync({ exp, runtime }), UPDATE_FREQUENCY);
   }
@@ -72,10 +75,27 @@ export class DevelopmentSession {
   }
 
   /** Stop notifying the Expo servers that the development session is running. */
-  public stop() {
+  public stopNotifying() {
     if (this.timeout) {
       clearTimeout(this.timeout);
     }
     this.timeout = null;
+  }
+
+  public async closeAsync(): Promise<void> {
+    this.stopNotifying();
+
+    const deviceIds = await this.getDeviceInstallationIdsAsync();
+
+    if (!(await isAuthenticatedAsync()) && !deviceIds?.length) {
+      return;
+    }
+
+    if (this.url) {
+      await closeDevelopmentSessionAsync({
+        url: this.url,
+        deviceIds,
+      });
+    }
   }
 }